+++ src/Components/Parsers/UnionKeywords.php
@@ -2,21 +2,21 @@
 
 declare(strict_types=1);
 
-namespace PhpMyAdmin\SqlParser\Components;
+namespace PhpMyAdmin\SqlParser\Components\Parsers;
 
-use PhpMyAdmin\SqlParser\Component;
 use PhpMyAdmin\SqlParser\Parseable;
 use PhpMyAdmin\SqlParser\Parser;
 use PhpMyAdmin\SqlParser\TokensList;
 use PhpMyAdmin\SqlParser\Translator;
 use RuntimeException;
+use Stringable;
 
 use function implode;
 
 /**
  * `UNION` keyword builder.
  */
-final class UnionKeyword implements Component, Parseable
+final class UnionKeywords implements Parseable
 {
     /**
      * Parses the tokens contained in the given list in the context of the given parser.
@@ -32,12 +32,7 @@ final class UnionKeyword implements Component, Parseable
         throw new RuntimeException(Translator::gettext('Not implemented yet.'));
     }
 
-    public function build(): string
-    {
-        throw new RuntimeException(Translator::gettext('Not implemented yet.'));
-    }
-
-    /** @param UnionKeyword[][] $component the component to be built */
+    /** @param list<array{string|Stringable,string|Stringable}> $component the component to be built */
     public static function buildAll(array $component): string
     {
         $tmp = [];
@@ -46,10 +41,5 @@ final class UnionKeyword implements Component, Parseable
         }
 
         return implode(' ', $tmp);
-    }
-
-    public function __toString(): string
-    {
-        return $this->build();
     }
 }
